i think this is how pr works
fix: made all handlers work properly 
added methods for uhhhh the stream_start and stuff and remove handlers work, and properly facilitated shutdown so that the plugin doesn't stack listeners when it shouldn't and also fixed the stream_start. two missing features exist. 1) if you panikk and unpanikk and then panikk again during a stream, it will turn on when you stop streaming, and i imagine the end user would not want it to turn itself on if they intentionally turned off powercord while streaming. 2) panikk happens when you set the keybind. i don't have a fix for that yet, though